Did it!
Managed to get everything working. The bug was in the calculation of u
and v using my built in functions: I was adding the freestream u
velocity and what I believed to be the doublet u velocity, but I was
calling the wrong variable for the doublet. Changed it, undid my random
fix attempts, and the plots are working. The drag value for part 6 still
isn't done though.