Passing sup_panels for the math
Trying to pass an array of all the panels (main and flap) for the math
so the sigma values are properly calibrated for no-normal flow at the
surface. For some reason, it's doing that fine but getting hung up on
the velocity calculations. Need to look at the function for calculating
the u and v for the flow.