Moving farther away to get closer
Before the gamma still seemed off. Really, I think it was the source
arrays (I was only including the effects of gamma on each airfoil when
creating the linear equations...meaning the two airfoils didn't talk to
each other!). Now the solution has blown up to huge values, but all of
the values have grown together so it seems like progress.